[17.0][MIG] web_view_leaflet_map: Refactor to OWL component
