+++ web_view_leaflet_map/hooks.py
@@ -4,15 +4,15 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 
-def uninstall_hook(cr, registry):
-    cr.execute(
+def uninstall_hook(env):
+    env.cr.execute(
         "UPDATE ir_act_window "
         "SET view_mode=replace(view_mode, ',leaflet_map', '')"
         "WHERE view_mode LIKE '%,leaflet_map%';"
     )
-    cr.execute(
+    env.cr.execute(
         "UPDATE ir_act_window "
         "SET view_mode=replace(view_mode, 'leaflet_map,', '')"
         "WHERE view_mode LIKE '%leaflet_map,%';"
     )
-    cr.execute("DELETE FROM ir_act_window " "WHERE view_mode = 'leaflet_map';")
+    env.cr.execute("DELETE FROM ir_act_window " "WHERE view_mode = 'leaflet_map';")
